+++ libs/containers/src/KeyValues_unittest.cpp
@@ -6,30 +6,26 @@
    | See: https://www.mrpt.org/Authors - All rights reserved.               |
    | Released under BSD License. See: https://www.mrpt.org/License          |
    +------------------------------------------------------------------------+ */
-#pragma once
 
+#include <gtest/gtest.h>
+#include <mrpt/config.h>
 #include <mrpt/containers/KeyValues.h>
 
-namespace mrpt::system
-{
-/** For passing arguments by name:
- *  \code
- *    TParameters<double> p;  // or TParametersDouble
- *    p["v_max"] = 1.0;  // Write
- *    ...
- *    cout << p["w_max"]; // Read, even if "p" was const.
- *  \endcode
- *
- * This is a type alias for mrpt::containers::KeyValue, kept for backwards
- * compatibility for MRPT<2.0.5.
- *
- * \deprecated (Since mrpt 2.0.5) Prefer mrpt::containers::Parameters.
- *
- * \ingroup mrpt_system_grp
- */
-template <typename T>
-using TParameters
-	[[deprecated("Prefer mrpt::containers::Parameters [Since mrpt 2.0.5]")]]  //
-	= mrpt::containers::KeyValues<T>;
+#include <algorithm>  // count()
 
-}  // namespace mrpt::system
+TEST(KeyValues, emptyCtor)
+{
+	mrpt::containers::KeyValues<double> kv;
+	EXPECT_TRUE(kv.empty());
+}
+
+TEST(KeyValues, SettersGetters)
+{
+	mrpt::containers::KeyValues<double> kv;
+	kv["foo"] = 1.0;
+	kv["bar"] = 2.0;
+
+	EXPECT_EQ(kv["foo"], 1.0);
+	EXPECT_EQ(kv["bar"], 2.0);
+	EXPECT_EQ(kv.getWithDefaultVal("chi", .0), 0.0);
+}