+++ libs/hwdrivers/src/CNTRIPClient.cpp
@@ -7,8 +7,6 @@
    | Released under BSD License. See: https://www.mrpt.org/License          |
    +------------------------------------------------------------------------+ */
 
-#include "hwdrivers-precomp.h"  // Precompiled headers
-
 #include <mrpt/comms/CClientTCPSocket.h>
 #include <mrpt/comms/net_utils.h>
 #include <mrpt/core/bits_math.h>
@@ -16,8 +14,11 @@
 #include <mrpt/hwdrivers/CNTRIPClient.h>
 #include <mrpt/math/wrap2pi.h>
 #include <mrpt/system/string_utils.h>
+
 #include <cstring>
 #include <iostream>
+
+#include "hwdrivers-precomp.h"	// Precompiled headers
 
 using namespace mrpt;
 using namespace mrpt::comms;
@@ -143,7 +144,7 @@ void CNTRIPClient::private_ntrip_thread()
 					// Nothing to be done... just wait
 				}
 
-				if (last_thread_do_process)  // Let the waiting caller continue
+				if (last_thread_do_process)	 // Let the waiting caller continue
 					// now.
 					m_sem_sock_closed.set_value();
 
@@ -325,7 +326,7 @@ bool CNTRIPClient::retrieveListOfMountpoints(
 {
 	string content;
 	int http_code;
-	TParameters<string> my_headers;
+	mrpt::containers::KeyValues<string> my_headers;
 
 	out_list.clear();