+++ libs/hwdrivers/src/CVelodyneScanner.cpp
@@ -7,8 +7,6 @@
    | Released under BSD License. See: https://www.mrpt.org/License          |
    +------------------------------------------------------------------------+ */
 
-#include "hwdrivers-precomp.h"  // Precompiled headers
-
 #include <mrpt/comms/net_utils.h>
 #include <mrpt/core/reverse_bytes.h>
 #include <mrpt/hwdrivers/CGPSInterface.h>
@@ -16,14 +14,17 @@
 #include <mrpt/serialization/CArchive.h>
 #include <mrpt/system/datetime.h>  // timeDifference
 #include <mrpt/system/filesystem.h>
+
 #include <thread>
+
+#include "hwdrivers-precomp.h"	// Precompiled headers
 
 // socket's hdrs:
 #ifdef _WIN32
 #define _WINSOCK_DEPRECATED_NO_WARNINGS
 #if defined(_WIN32_WINNT) && (_WIN32_WINNT < 0x600)
 #undef _WIN32_WINNT
-#define _WIN32_WINNT 0x600  // Minimum: Windows Vista (required to pollfd)
+#define _WIN32_WINNT 0x600	// Minimum: Windows Vista (required to pollfd)
 #endif
 
 #include <winsock2.h>
@@ -46,12 +47,13 @@ using socklen_t = int;
 #include <sys/time.h>  // gettimeofday()
 #include <sys/types.h>
 #include <unistd.h>
+
 #include <cerrno>
 #endif
 
 #if MRPT_HAS_LIBPCAP
 #include <pcap.h>
-#if !defined(PCAP_NETMASK_UNKNOWN)  // for older pcap versions
+#if !defined(PCAP_NETMASK_UNKNOWN)	// for older pcap versions
 #define PCAP_NETMASK_UNKNOWN 0xffffffff
 #endif
 #endif
@@ -503,7 +505,7 @@ void CVelodyneScanner::initialize()
 				filter_str += "&& src host " + m_device_ip;
 
 			static std::string sMsgError =
-				"[CVelodyneScanner] Error calling pcap_compile: ";  // This is
+				"[CVelodyneScanner] Error calling pcap_compile: ";	// This is
 			// to avoid
 			// the
 			// ill-formed
@@ -585,16 +587,16 @@ void CVelodyneScanner::close()
 
 // Fixed Ethernet headers for PCAP capture --------
 #if MRPT_HAS_LIBPCAP
-const uint16_t LidarPacketHeader[21] = {0xffff, 0xffff, 0xffff, 0x7660, 0x0088,
-										0x0000, 0x0008, 0x0045, 0xd204, 0x0000,
-										0x0040, 0x11ff, 0xaab4, 0xa8c0, 0xc801,
-										0xffff,  // checksum 0xa9b4 //source ip
-										// 0xa8c0, 0xc801 is
-										// 192.168.1.200
-										0xffff, 0x4009, 0x4009, 0xbe04, 0x0000};
+const uint16_t LidarPacketHeader[21] = {
+	0xffff, 0xffff, 0xffff, 0x7660, 0x0088, 0x0000, 0x0008, 0x0045, 0xd204,
+	0x0000, 0x0040, 0x11ff, 0xaab4, 0xa8c0, 0xc801,
+	0xffff,	 // checksum 0xa9b4 //source ip
+	// 0xa8c0, 0xc801 is
+	// 192.168.1.200
+	0xffff, 0x4009, 0x4009, 0xbe04, 0x0000};
 const uint16_t PositionPacketHeader[21] = {
 	0xffff, 0xffff, 0xffff, 0x7660, 0x0088, 0x0000, 0x0008, 0x0045, 0xd204,
-	0x0000, 0x0040, 0x11ff, 0xaab4, 0xa8c0, 0xc801, 0xffff,  // checksum 0xa9b4
+	0x0000, 0x0040, 0x11ff, 0xaab4, 0xa8c0, 0xc801, 0xffff,	 // checksum 0xa9b4
 	// //source ip
 	// 0xa8c0, 0xc801
 	// is 192.168.1.200
@@ -778,7 +780,7 @@ mrpt::system::TTimeStamp CVelodyneScanner::internal_receive_UDP_packet(
 	struct pollfd fds[1];
 	fds[0].fd = hSocket;
 	fds[0].events = POLLIN;
-	static const int POLL_TIMEOUT = 1;  // (ms)
+	static const int POLL_TIMEOUT = 1;	// (ms)
 
 	sockaddr_in sender_address;
 	socklen_t sender_address_len = sizeof(sender_address);
@@ -813,7 +815,7 @@ mrpt::system::TTimeStamp CVelodyneScanner::internal_receive_UDP_packet(
 				WSAPoll
 #endif
 				(fds, 1, POLL_TIMEOUT);
-			if (retval < 0)  // poll() error?
+			if (retval < 0)	 // poll() error?
 			{
 				if (errno != EINTR)
 					THROW_EXCEPTION(format(
@@ -851,7 +853,7 @@ mrpt::system::TTimeStamp CVelodyneScanner::internal_receive_UDP_packet(
 				sender_address.sin_addr.s_addr != devip_addr)
 				continue;
 			else
-				break;  // done
+				break;	// done
 		}
 
 		std::cerr
@@ -918,7 +920,7 @@ bool CVelodyneScanner::internal_read_PCAP_packet(
 				memcpy(
 					out_pos_buffer, pkt_data + 42,
 					CObservationVelodyneScan::POS_PACKET_SIZE);
-				pos_pkt_time = tim;  // success
+				pos_pkt_time = tim;	 // success
 				return true;
 			}
 			else if (udp_dst_port == CVelodyneScanner::VELODYNE_DATA_UDP_PORT)
@@ -942,7 +944,7 @@ bool CVelodyneScanner::internal_read_PCAP_packet(
 			}
 		}
 
-		if (m_pcap_file_empty)  // no data in file?
+		if (m_pcap_file_empty)	// no data in file?
 		{
 			fprintf(
 				stderr,
@@ -1078,7 +1080,7 @@ bool CVelodyneScanner::internal_send_http_post(const std::string& post_data)
 	string post_err_str;
 
 	int http_rep_code;
-	mrpt::system::TParameters<string> extra_headers, out_headers;
+	mrpt::containers::KeyValues<std::string> extra_headers, out_headers;
 
 	extra_headers["Origin"] = mrpt::format("http://%s", m_device_ip.c_str());
 	extra_headers["Referer"] = mrpt::format("http://%s", m_device_ip.c_str());
@@ -1092,7 +1094,7 @@ bool CVelodyneScanner::internal_send_http_post(const std::string& post_data)
 		&http_rep_code, &extra_headers, &out_headers);
 
 	return mrpt::comms::net::erOk == ret &&
-		   (http_rep_code == 200 || http_rep_code == 204);  // OK codes
+		   (http_rep_code == 200 || http_rep_code == 204);	// OK codes
 
 	MRPT_END
 }